BE: Fix Tests (because of extension of Route model)
