+++ backend/src/test/java/de/neuefische/backend/service/RouteServiceTest.java
@@ -18,8 +18,10 @@ class RouteServiceTest {
     private final IdService idService = mock(IdService.class);
     private final LocationService locationService = mock(LocationService.class);
 
+    private final RoutesService routesService = mock(RoutesService.class);
+
     private final RouteService routeService = new RouteService(routeRepository,
-            idService, locationService);
+            idService, locationService, routesService);
 
     @Test
     void addNewRoute_ShouldReturn_AddedRoute(){
@@ -27,8 +29,10 @@ class RouteServiceTest {
         String[] hashtags = new String[1];
         hashtags[0] = "tree";
         StartPosition startPosition = new StartPosition(2.2, 1.1);
-        RouteDTO routeDTO = new RouteDTO("routeName", hashtags, "imageThumbnail", startPosition);
+        EndPosition endPosition = new EndPosition(2.3, 1.12);
+        RouteDTO routeDTO = new RouteDTO("routeName", hashtags, "imageThumbnail", startPosition, endPosition);
         when(idService.generateId()).thenReturn("1");
+        when(routesService.getRoutes(any(), any())).thenReturn(null);
         when(routeRepository.save(any())).thenReturn(
                 Route.builder()
                         .id("1")
@@ -36,6 +40,8 @@ class RouteServiceTest {
                         .hashtags(routeDTO.getHashtags())
                         .imageThumbnail(routeDTO.getImageThumbnail())
                         .startPosition(routeDTO.getStartPosition())
+                        .endPosition(routeDTO.getEndPosition())
+                        .routes(null)
                         .position(new GeoJsonPoint(2.2, 1.1))
                         .build()
         );
@@ -49,6 +55,8 @@ class RouteServiceTest {
                 .hashtags(routeDTO.getHashtags())
                 .imageThumbnail(routeDTO.getImageThumbnail())
                 .startPosition(routeDTO.getStartPosition())
+                .endPosition(routeDTO.getEndPosition())
+                .routes(null)
                 .position(new GeoJsonPoint(2.2, 1.1))
                 .build();
         assertEquals(expected, actual);
